"7:15pm. It is this late? Damn.
At any rate, I've figured out pretty much everything. I know how to do eager let generalization now. I also know how to deal with constraints during the forward pass. I won't be able to deal with all of them during the first inference, but nonetheless, the way I am handling `RecordApply` right now is abysmal. There is a much better way.

I also know how to deal with forall vars without that ugly hack.

Also...I am going to change the constraint syntax compared to what I had in mind. I have a different perspective on them now.

7:20pm. I'd go into detail, but since I still haven't had lunch, let me get to that first.

8:30pm. Done with lunch. I do not feel like talking about it right now.

...

Let me do it.

///

I figured it out. The way to propagate scope information succinctly is to adjust unification is done.

The way I am doing it right now, is to unify metavariables with compound types directly.

```
inl f x =
    inl g forall a. (y : a) =
        (y, 1) = x
    ()
```

For unifying `(y, 1) = x`, I would do something like the following. `x` would be a metavariable, so I would unify `'0 => (a, i32)`. Right now I am just using unique ids to differentiate between metavariables and have no scope information. But supposing I had scope information, what I could do is.

`'0 (0) => (a (1), i32)`

I had this idea before, but discarded it because that is still not enough to make this an error. Strictly speaking, in this example there is, but suppose I had the following chain of unifications...

`'0 (0) => ('1 (1), '2 (1)) => (a (1), i32)`, then that would break.

This chain would correspond to code fragment like...

```
inl f x =
    inl g forall a. (y : a) =
        inl q,w = x
        (y, 1) = (q,w)
    ()
```

The way to properly unify this is to do something counter intuitive. Even though at the point this equality constraint is being processed the scope is `1`, what I need to do is make a bunch of dummy metavariables with the previous scope of `0`.

Then there is a rule if I am unifying two metavariables, the arrow points in the direction of the lower scope.

If I had to unify something like `('2 (1), '4 (0)) => ('0 (0), '5 (3))`, those would reduce to `'2 (1) => '0 (0)` and `'5 (3) => '4 (0)`.

If I had to unify something like `'0 (0) => ('1 (1), '2 (1)) => (a (1), i32)`, it would be better to first extend it to `'0 (0) => ('3 (0), '4 (0)) => ('1 (1), '2 (1)) => (a (1), i32)`

Since `'1 (1) => '3 (0)` and `'2 (1) => '4 (0)`, that would leave unifying `('3 (0), '4 (0)) => (a (1), i32)`. Then `'3 (0) => a (1)` is an obvious scope error. The great thing is that this kind of unification would propagate scope information in an optimal manner, so if I the type of statement is `'5 (1) -> '5 (1)` and the current scope is `1`, I know I can generalize this. On the other hand if it is `'5 (0) -> '5 (0)`, then I cannot. And thanks to this I should be able to do eager let generalization.

It seems simple, but the way I have the unifier implemented currently has made me myopic to this kind of structure. It is my first time doing top down type inference, so I keep missing obvious things. One thing I can't really use is separating equality constraints from the overall program as [SPJ suggests](https://www.youtube.com/watch?v=x3evzO8O9e8). Type inference needs to be flow sensitive to an extent in order to deal with records and separating unifyication of equalities from the overall program would break that.

One type inference problem that is still difficult to me is how to deal with propagating constraints in mutually recursive blocks.

///

9:10pm. Let me also check out the answer I got on the PL sub."