"3:40pm. I am still thinking. I guess today will pass like that.
I've realized something.

```
inl f forall a b c. ....
```

```fs
        | RawForall(r,(_,(name,k)),b) ->
            let rec typevar = function
                | RawKindWildcard -> fresh_kind()
                | RawKindStar -> KindStar
                | RawKindFun(a,b) -> KindFun(typevar a, typevar b)
            let k = typevar k
            let i,v = fresh_var'' k
            forall_scopes.Add(name,i)
            let body = fresh_var()
            unify r s (TyForall((name,k),body))
            term {env with ty = Map.add name v env.ty} body b
```

Consider what the scopes in this would be.

Each nested forall would have its own scope.

Since I am propagating info forward in Spiral, that means I will end up unifying `'0 (0) => forall a. '1 (1)`. Given my current rules, that would end up unfying `'1 (1)` with something that has scope of 0.

The solution is for each statements to take stock of the foralls at the outset and start inference at a more sensible place.

3:50pm. So it is not that I can just put forall whenever I want. For both ordinary and recursive statements I already need some special behavior because infering foralls is ill behaved normally.

For mutually recursive statements...yeah, in the presence of a single forall they have to be fully annotated. Otherwise they do not.

Now that I am thinking about scopes, the interesting thing about unifying a metavariable with an expression body is that it should always have scope that is at most the current scope.

Oh yeah, when it comes to dependent constraints, the metavariable's scope will have to have a minimum scope of the variables in its constraints.

...No that is the wrong way to put it. No actually, what I need to say is that the constraint's variables need have at most the scope of the metavariable itself.

When unifying metavariables, I need to unify the scope of its constraints as well.

4pm. At any rate, what makes this whole thing possible is that foralls have explicitly stated constraints. If my intention was to fully infer those as well, what I have in mind would not work and I would require multiple passes in order to converge to all the constraints.

4:10pm. Thunder. Let me run here."