"```fs | TyMetavar of Var * T option ref ```
Let me go back to this. Yeah, I know that the other is more succinct, but it would be a pain in the ass to pattern match on.

2:25pm. Let me finish reading the chapter of the shitty Chinese LN I am hooken on currently and I will resume the second TAPL book.

2:45pm. Ok, let me resume.

To be honest, my obsession has quieted down, and I feel like taking a nap now rather that do anything else, but let me instead go through the book. I should at least put two hours into it today.

I think tomorrow, I should get back to programming.

3pm. 418/589. Not only is this book putting me to sleep, but I am also hearing thunder outside despite the sky being clear.

I am wasting my time here.

3:10pm. No forget it. I see the book recommended often on the PL sub, but it is a pile of shit. I am actually better off just thinking about the problem and trying to solve it myself instead of trying to interpret the book.

In programming, writing the code is actually easier than the converse of it. The same goes for math.

You have a mental model, you turn it into code. The reverse is much harder.

Probably the greatest example of this is nature itself - a brainless process like it produced such marvels of engineering, but reverse engineering it is a very difficult task. Just where did the understanding go?

These math books fall into the same bucket.

Though I did not get what I want from my math studies in 2019, I did benefit greatly in that I now understand how to categorize math. I also know what proofs are.

Mathematicians are really strange beasts. In constrast to programmers whose primary concern is communicating with the machine and nature, they are more social with respect to people.

There has been a rift. I think many people who would have become great mathematicians in the past would be programmers today. It is not appreciated that this shift would change the character of the whole mathematics community itself.

3:20pm. Let me step away from the screen for a while. I'll take a nap, cultivate my desire and then just start.

It does not really matter if the TC takes a month or three. What matters is that I have confidence in my own work. 2020 is that kind of year.

I'll embrace my loneliness and have it strengthen me.

This is the final stretch. After this redesign, Spiral's type system will be complete. Not half finished like I imagined it would be originally, but completely done. I won't do this, move on the later stages, only to have to come back to put in constraints or other things.

The testing when it starts will be over the whole thing. And that is a huge benefit.

If I can grasp it, I should do so by any means possible.

I need to focus on this desire. I need to imagine how great it will be to program in Spiral v0.2."