"7:20pm. I've fortified my mind somewhat.
This whole lethargy feels like it is socially induced for me. It is my weakness. When my energy falls, always that injury comes to the front of my mind. It is loneliness.

On paper I have friends, but suppose I did something like write a book, or write a programming language, or figure out the meaning of life. No matter how I look at it, I cannot imagine getting any respect for it and in fact, I did not. Sure making a million dollars would impress them, but that would impress anybody. What good are 'friends' when their only talent is at irony? They are untrustworthy idiots who will look for the meaning of life at the bottom of a wine bottle until they die.

The work I am doing on Spiral here is impressive. The way I've figured out how to do type inference pretty much all on my own takes some skill. Sure, it might not impress anybody in my circle, but I can certainly imagine people in this world finding value in Spiral. Especially v0.2 when it comes out. It will be a powerful bargaining chip too.

Friendship has a purpose. In the past friends would have been hunting partners. The contexts would have been such that if one does not perform then both starve.

This is what I really crave for. I am going to finish Spiral and then use it to find friendship. The people driving the transition from personal computer to personal supercomputing will need it. It won't be close and personal, but based on the last 20 or so years, I would have been better off avoiding friendship where there are no shared interests or goals between the parties.

I want to trade. And I want to run towards the same goal with others. I want to associate with people who would prefer to elevate their intelligence rather than suppress it.

7:35pm. Consider the context. It is not whether one has social skills or does not. Despite similar outward appearance, two people can be different species after all. I can't be friendly and outgoing after the other side spends hours and hours boring me to death.

Maybe the problem is that one is simply at the wrong place and at the wrong time, amongst wrong people. When the way forward is clear, the feeling will come. I won't make this mistake again.

Spiral is going to be impressive. And I am going to find people worth impressing with it. And in the background the story will slowly move closer to the Singularity, by those who would later become the Inspired. The real ones, not just from a piece of fiction.

7:40pm. The work on Spiral is worth doing. Unless I do this, there will be shit tier languages dominating the scene until somebody comes along and does what I should have done.

I won't let it come to that. I was too late for Cuda, but I can catch the next train.

I need to get that 'friendship as power sharing' thing going on. This is not what frienship was for me and I really regret it. If it could have met this standard, I would have been a lot happier person throughout my existence.

7:45pm. I am just sad. I keep moving forward, but the rest aren't following.

Tomorrow, I will really endeavor to write some code. I've thought long and hard about what I should do, and this is an integral part of programming, but now it is time that I get some carving done as well.

I need to get into it again. Stopping for so long after doing that useless record apply thing has distrupted my moment and I need to go through it from the top.

The type inferencer is possibly one of my greatest pieces of work. Right now it is half finished, and I still feel it came out great. It is amazing how good the work is considering I barely have any experience in this subject.

The code in `pattern`, `term` and `type` will not have to go through much change at all - the real changes will all be in the unifier. I do not have that much work to go through at all.

7:50pm. There is something I regret. I bash Chinese sometimes for their ancient medicine bullshit, but apart from that, cultivators are wizards done right.

One thing I regret buying into is the short termism of the normies. When I started six months was a long time, but assuming I have my normal lifespan left to live out, that is a huge time left for me to do programming.

I should aim to become a proper wizard.

Programming is really the only avenue one can take to surpass one's natural limits in reality. I should take that thought in stride since I will be mostly doing janitorial work in the future.

7:55pm. Even after I get those chips, no doubt I will get tired of managing memory, so at that point I'll rope my sponsor into paying me to work on GC. The way I am going, everything is going to be covered until I get my true power.

But if I am going to get power, I should first get dignity. And to do that I need to get rid of the parasites squatting rent free in my mental room, and find some better tenants.

And to do that I need to start really believing in Spiral. I've always been driven by my sense of aesthetics towards it. That this is the way I see my own programming is really great. It is a lot better than it being some meaningless job. But it is possible to do even better.

Spiral should not be a power of my future, but the power of my present as well.

People go to colleges for years in order to pad their resumes.

Me, I will just point them at the typechecker and watch them stand back in awe.

8pm. Yes, I want that top down type system. But the way the typechecker itself is written is worthy of adulation.

I need to find the desire to program more. The careful, painstaking attention to detail of the past two weeks, and then the recent raging brainstorming is heroic. I should continue this kind of unrecognized heroism.

8:05pm. One day my talent will blossom, and the world will be brought to heel. If peace can only be brought through power, then truth will have to be brought through treachery. Those on the other side deserve what is coming to them...and so do we deserve what is coming to us.

I want to live in a world where effort pays off. I want to live in a world where those like me can find the power that they seek. I cannot stop. I have to act in order to fulfill that desire."