"1:55pm. Just a bit more and I will start.
2:20pm. Let me start. From here on out, I will aim to refocus on the task at hand. The amount of work left to be done is intimidating, but I will cover it.

After `subst` the next function is `infer`.

```fs
let infer (aux : AuxEnv) (top_env : Env) (env : Env) x : T =
    let errors = ResizeArray()
    let term' = ResizeArray()
    let kind = ResizeArray()
    let forall_scopes = System.Collections.Generic.Dictionary()
```

That `forall_scope` dictionary always sat poorly by me, and I was not in thinking that wrong. I also won't need those arrays for terms and kinds.

When I studied unification in 2017, the Haskell and the ML versions used a map in order to store ids. At least, I think that was how it was, so I just based it on my memory.

This scheme would work, but clearly just having a reference directly in the metavariable is far superior as a scheme.

```fs
let fresh_kind () = let x = KindMetavar kind.Count in kind.Add(x); x
```

I do not need to keep track of counts for the kinds anymore. The same goes for regular vars.

```fs
let fresh_kind () = KindMetavar (ref None)
```

This is all I need.

```fs
let fresh_var x = TyMetavar (x, ref None)
```

Let me make it like so for now.

```fs
    let rec kind_get i =
        match kind.[i] with
        | KindMetavar i' as x -> if i <> i' then let x = kind_get i' in kind.[i] <- x; x else x
        | x -> kind_subst x
    and kind_subst = function
        | KindMetavar i -> kind_get i
        | KindFun(a,b) -> KindFun(kind_subst a,kind_subst b)
        | KindStar -> KindStar
```

Let me adjust this.

```fs
    let rec kind_subst = function
        | KindMetavar ({contents=Some x} & link) -> let x = kind_subst x in link := Some x; x
        | KindFun(a,b) -> KindFun(kind_subst a,kind_subst b)
        | x -> x
```

Yeah, this should be like so. I know I did those visit functions, but this does in depth substitution. I am not sure whether I want this, but it will be convenient when it comes to doing the occurs check and such. For regular unification I would trigger this.

```fs
        | TyRecordApply(a,b) ->
            match f a, f b with
            | TyRecord l, TySymbol x ->
                match Map.tryFind x l with
                | Some x -> x
                | None -> errors.Add(r,RecordIndexFailed x); fresh_var()
            | (TyRecord _ | TyVar _ | TyMetavar _) & a, (TySymbol _ | TyVar _ | TyMetavar _) & b -> TyRecordApply(a,b)
            | (TyRecord _ | TyVar _ | TyMetavar _), b -> errors.Add(r,ExpectedSymbolAsRecordKey b); fresh_var()
            | a,_ -> errors.Add(r,ExpectedRecord a); fresh_var()
```

Let me get this thing out. This should come in the constraint somewhere.

```fs
    let rec term_subst r x =
        let f = term_subst r
```

Ah, yes. I went adding the range for this everywhere.

```fs
    let rec term_subst r x =
        let f = term_subst r
        match x with
        | TyConstraint _ | TyVar _ | TyHigherOrder _ | TyB | TyPrim _ | TySymbol _ as x -> x
        | TyPair(a,b) -> TyPair(f a, f b)
        | TyRecord l -> TyRecord(Map.map (fun _ -> f) l)
        | TyFun(a,b) -> TyFun(f a, f b)
        | TyForall(a,b) -> TyForall(a,f b)
        | TyArray a -> TyArray(f a)
        | TyApply(a,b,c) -> TyApply(f a, f b, c)
        | TyInl(a,b) -> TyInl(a,f b)
        | TyMetavar(i,{contents=Some x} & link) -> let x = f x in link := Some x; x
        | TyMetavar((_,cons,_,_) & i, link) -> eval_constraints cons x
```

Oh, shit. It is only now that I've come this far, that I really understand in how much shit I am.

Oh crap. How is the set supposed to compare variables when they do not have unique ids?

Even worse. It completely slipped my mind that forall vars need to be fully specified ahead of time. What if its constraints are metavariables?

I did not think about this at all.

3:15pm. Oh crap. I am making a breakthrough in my understanding of constraints, but all this is making me realize is just how impossible the record apply is.

I can't typecheck it. Even if the metavar ids in constraints or metavars in constraints in foralls weren't an issue, it would still be too difficult. Like, with nested functions, the constraints would not be enough to capture what I need.

3:20pm. No, it is not that it can't be done. It is that `record_apply` as a constraint does not make sense. Rather what I need is a partial record constraint.

Something like

```
forall x r {record_field .a x}.
```

Instead of having the constraint be on the result of record apply, it needs to be on the record itself. Then it would work.

An alternative way of doing it that is more generic.

```
forall x s {symbol} r {record_field s x}.
```

Using this signature I could make the record apply function.

3:25pm. That having said, there is still the can of worms that are metavariables in constraints.

3:30pm. No, not even this would work.

3:35pm. Ok, as much as I want dependent constraints in order to be able to do record application at the type level, I will abandon the idea if it turns out to be too unworkable.

Let me step away from the screen for a while. I really thought I would do some significant programming, but it seems I am going back to brainstorming.

It is like I can only think so far into the future without actually sitting down and doing.

The hand does the office of the mind. This is an example of that saying in action.

At worst, maybe I will just do single parameter constraints. That would be useful and really easy to deal with."