"5:05pm. Done with both lunch and thinking.
I've come to a decision. All dependent constraints are out. That means no `record_field` or `record_apply`.

Furthermore, prototype constraints won't be dependent.

```
prototype qwe a b c = ...
inl f {qwe} ...
```

Only the first parameter will have constraints associated with it. The inner constraints, like in list equality from Wadler's paper will only apply to the first parameter. That will simplify thing with regard to prototypes. In all respects they will be single parameter typeclasses.

Now, I've decided what I can use to substitute record application. I will do a split function whose signature is:

`forall (f : * -> *) a. tensor (f a) -> f (tensor a)`

The way it will be used is that I will make some nominal like...

```
nominal dim t = {a : t; b : t; c : t}
inl g _ =
    inl d = dim {a=1; b=2; c=3}
    inl {a b c} = tensor_init dim |> split
    a ! 0 + b ! 0 + c ! 0
```

This will do nicely. The important thing is not that I can apply dimensions, but that I can split them. Of course, you can always write something erroneous like...

```
nominal dim t = {a : t; b : t; c : i32}
inl g _ =
    inl d = dim {a=1; b=2; c=3}
    inl {a b c} = tensor_init dim |> split
    a ! 0 + b ! 0 + c ! 0
```

I'll have to do some checking in the `split` function to make sure that the body of `f a` is a record and that the result matches what one would expect to come out based on the signature. But that is an easy job for the partial evaluator.

This kind of scheme is an acceptable compromise.

```fs
let id_to_kind x =
    let (^) a b = KindFun(a,b) // ^ is right assoc
    match x with
    | CINumber -> KindType ^ KindConstraint
    | CIRecordApply -> KindType ^ KindType ^ KindType ^ KindConstraint

type Constraint =
    | CNumber
    | CRecordApply of T * T
```

Let me get rid of record apply. Constraints should be simple after all. All the constrains from here on out will only need to be evaluated when the metavar unifies with a principal type. One thing that really stressed me out is that I would have needed to constantly go through the constraint set otherwise in the term subst. That was just nasty.

What I have in mind is just too hard.

```fs
type ConstraintId =
    | CINumber // * -> /

let id_to_kind x =
    let (^) a b = KindFun(a,b) // ^ is right assoc
    match x with
    | CINumber -> KindType ^ KindConstraint

type Constraint =
    | CNumber
```

So it is just number for now. That is acceptable.

```fs
    let rec term_subst x =
        match x with
        | TyMetavar(i,{contents=Some x} & link) -> let x = term_subst x in link := Some x; x
        | TyMetavar _ | TyConstraint _ | TyVar _ | TyHigherOrder _ | TyB | TyPrim _ | TySymbol _ as x -> x
        | TyPair(a,b) -> TyPair(term_subst a, term_subst b)
        | TyRecord l -> TyRecord(Map.map (fun _ -> term_subst) l)
        | TyFun(a,b) -> TyFun(term_subst a, term_subst b)
        | TyForall(a,b) -> TyForall(a,term_subst b)
        | TyArray a -> TyArray(term_subst a)
        | TyApply(a,b,c) -> TyApply(term_subst a, term_subst b, c)
        | TyInl(a,b) -> TyInl(a,term_subst b)
```

I'll have to get rid of the ranges I am passing into the call to `term_subst` when I refactored it for record apply.

But I'll replace a bunch of those calls with `visit_t` anyway. I'll refactor it all in turn. It is fine.

5:35pm. Now, I am actually pretty tired so I will call it a day here. I do not feel like programming anymore.

5:40pm. At any rate, now that I finally resolved the whole record apply issue by replacing it with `split` and decided to kick out dependent constraints, things will be much easier for me. At this point the type system design feels completely coherent. I should not have any surprises tomorrow.

5:45pm. Right now I am thinking whether I want to allow nominal and union's type parameters to have constraints.

It might be useful. Well, I'll leave this feature for later. It would not be hard to add either way.

It is not the kind of thing that would give me lack of sleep like dependent constraints.

5:55pm. I am still thinking about it. I need to reinitialize myself anyway so let me close here. Today I wanted to start, but my pace was distrupted. Tomorrow I will get this thing going."