"4:40pm. No those good ideas are nothing. They would introduce cycles.
So the only way to multi arg constraints is to have an array and do all processing over it.

For once, I think I am going to say that this is too much. If I really want my record apply, I can have it, but I better be prepared to pay up. And to be honest, I am not. I do not want to pay up.

The way the inference is implemented now is great, and multi arg constraints would wreck it.

4:45pm. It is just for higher ranked types, map constraints and various other ideas. You can sit down and put them in, and they would work, but at certain point you need to draw the line between what are essential and non-essential features in a language. Especially when there are decent substitutes for the non-essential features.

4:50pm. Do I want to work until I have the perfect language, or do I want I want finish this by the end of the month and go to the next phase?

Finish, finish, finish!

I'll take care of v0.2, and the agents will take care of the rest.

4:55pm. I am not competing with Haskell or Scala or Rust here on which language has the most elaborate type system.

I do not even need the prototype constraints. They would take too much time. In fact, they would make the language less flexible. Literally every use of typeclasses I can think of can have a bottom-up default.

Forget the complicated things. Finish the language. It is more important that I make that next version of the ML library in Spiral v0.2. It is more important that I get my hands on those next gen chips.

I am not trying to impress other people with a bullet list of features. Nobody is going to give Spiral a pass if typechecking of prototypes is deferred to the partial evaluation stage.

...Though they might give it a pass if I at least do not have the number constraint. That one clearly needs to be in.

It just goes to show how the simple things are the most important.

5pm. I am really glad that I took time to do the most basic of the constraints. That was really great of me. Being able to tell the type system that forall is constrained to a number will pay off.

5:05pm. Now, although it is not interesting as multi arg constraints and other fancy features, I need to focus on on finally dealing with those damn lets.

```fs
        //| RawMatch of Range * body: RawExpr * (Pattern * RawExpr) list
        //| RawRecBlock of Range * ((Range * VarString) * RawExpr) list * on_succ: RawExpr
```

Tomorrow, I will imagine that the gun pointing at me, and start seriously focusing on these cases. I have what it takes to do them now. I went thought them over and over in my head way too many times. I need to do these.

5:10pm. Right now isn't the time to start anything new though. I've been at it from an early point today and deserve my rest.

5:15pm. Tomorrow, I will face my enemy. These two cases I seem to be perenially stuck on, but tomorrow I will overcome them. Then I will be 'done' with the typechecker. I think tying up the loose ends will take a while after that, at least a few days. Probably more. But I should be able to move to testing after that.

I think that if I stick to my decision of not adding any new features, I have a really good chance of being completely done with the typechecker by the end of the month. Hopefully I will have gotten type on hover and documentation parts of editor support by that time as well. But if I can't then it is no biggie. I'll definitely be able to deal with it sometime early next month in that case.

5:20pm. Let me try to unwind here."