"5:25pm. https://www.reddit.com/r/ProgrammingLanguages/comments/i9al92/how_to_market_haskell_to_a_mainstream_programmer/
Let me watch this for a bit.

6pm.

```fs
| TyMetavar(V(scope',q,w,e),_) when scope < scope' -> fresh_var'' (V(scope,q,w,e))
```

Ah, I screwed this up.  I need to connect the new variable to the old one.

```fs
| TyMetavar(V(scope',q,w,e),link) when scope < scope' -> let x = fresh_var'' (V(scope,q,w,e)) in link := Some x; x
```

Yeah, this is why I do not want to bloat the implementation any more than it needs to be. The `number` constraint is enough for me. Maybe it will turn out to be quite a while before I put in anything else.

If one does not put a leash on the program complexity, it can easily baloon out of control. The sure way to have it explode in me is to start putting in things I do not need, but might. Once you start thinking like that, you start dying a slow death as programmer."