"5:10pm. Yeah, telling a lie to the typechecker is the right thing to do here.
Suppose instead of indexing into a record, I wanted to unzip a tensor.

`forall a el el'. tensor a el -> (el -> el') -> tensor a el'`

Then this is the right thing to do. I would need the same technique again. Fancier constraints would not help me here.

Under the hood, the tensor stores each element in a separate array, but they are neatly ordered. A function that takes `el -> el'` would also work the same thing mapped to arrays. I would just need to make sure everything in `el'` exists in `el` and it would work.

I am sure of it.

The lie I would be doing here is more egregious than in other places, but it is the right thing to do. It is not truth, but utility and yet I cannot afford to pass over it. This is the right way to use the type system.

This technique cements the lesson that a type system is a tool and not a master.

5:20pm. Ok, I think I know how to make comparing many elements for membership efficient. I'll use a reference set and a special Op that takes a ()-terminated tuple rather than a single elements when checking for existence.

...Damn it, it won't work. I can't use tuples as pairs the same way I could have used tuples as lists back in v0.09.

Ah, wait, rather than limit the tuple with a (), I'll just use `elem: x` pairs. That way I won't dive too deeply accidentally.

An alternative would also be to just use a counted record - just number its keys as numbers. Either way is fine.

5:25pm. Then I will just put all the references in a HashSet and iteratively compare them. In such a situation I would want to avoid having to do a linear search for every element.

...Ok, let me take a bath here and I will call it a day. I still have to think about it, but I am warming up to this. Hopefully, this idea will give me enough confidence to proceed with the plan. Even today I still haven't gotten to the let statements."