"6:25pm. So another day will go like this, with me obsessing about things. Yeah, the extractor functions are the way to go. There are various ways of implementing them depending on the context.
They are an important technique. Their birth is the marriage of type inference and partial evaluation. They the herald of v0.2.

The feeling is good. I do not feel like dealing with multi arg constraints at all now.

I do not have to talk myself out of dealing with their hardship, they simply aren't as good as the extractor function technique. Some users might say that this is not true, but type inference being 10-100x faster should be enough to shut any complaints down. Doing a linear search over an array plus deduplication requiring `term_subst`ing everything would kill performance.

The only reason why unification itself can be performant is because of metavariable links. It is trully a wondorous technique.

Spiral is not like Haskell or F#. If it were those languages, they can afford to wait till the statement is done being typchecked before solving constraints. Spiral would need to do that after literally every unification if I go down that route. It would just be crazy.

It is all because of the nature of records. It probably is not an accident that they are so shitty in Haskell. Since all record accessors are functions in it, that gives it a lot more leeway and the typechecker can afford to be flow insensitive.

In Spiral, I overload application with a special case just for them.

6:40pm. Actually, I had not thought about this - the record with pattern. Instead of it requiring a record right away, could compile it down to constraints?

...It might be possible. But it would be complicated.

Still, suppose I succeeded in making the language flow insensitive, using complex constraints for everything would no doubt crush error message quality.

The extractor functions I will be using have the great advantage of propagating information directly.

6:45pm. Yes, it is decided. The language is perfect as it is. Tomorrow, I am going to do another redesign just to tighten up those foralls. I need to do another pass because I was too haphazard.

Then I will turn my attention to those let statements. I want to put the typechecking phase behind me and get to the meat of the language. Once I am done with this, I will be able to have a breath of fresh air. No more obsessing, no more difficult language design decisions.

I feel that once I deal with the typechecking phase, Spiral will have a more professional air about it."