"9:05am. I went to bed too late last night and I've been super groggy for the last 20m.
https://arxiv.org/abs/2007.05558
The Computational Limits of Deep Learning

> Deep learning's recent history has been one of achievement: from triumphing over humans in the game of Go to world-leading performance in image recognition, voice recognition, translation, and other tasks. But this progress has come with a voracious appetite for computing power. This article reports on the computational demands of Deep Learning applications in five prominent application areas and shows that progress in all five is strongly reliant on increases in computing power. Extrapolating forward this reliance reveals that progress along current lines is rapidly becoming economically, technically, and environmentally unsustainable. Thus, continued progress in these applications will require dramatically more computationally-efficient methods, which will either have to come from changes to deep learning or from moving to other machine learning methods.

I think I'll take a look at this.

9:10am. Yesterday I hit the milestone, so I am going to take it easy today for a bit.

9:40am. Let me slack more. That paper I linked to is too long winded and is just wasting my time.

9:45am. I've really gotten into Against The Gods yesterday, so let me have that. Then I'll start programming.

10:05am. The translation quality really feel of a cliff in today's session. Forget this shit.

Let me read that extra 20.5 Knight Run chapter and then I will start.

10:25am. That was fun. But let me finally start. I need to get into it for a while.

Spiral won't get done on its own. I need to be the one to push it.

10:30am. Focus me.

10:35am. I've indulged myself a good bit. Now it is time to get serious. I allowed myself to slacken the pressure, but now it is time to raise it again.

First of all, I've decided - I am going to leave out filling the AST until it is time to deal with the prepass. Forget about that thing until then.

10:40am. Ok, prototype instances are one thing. After I finish that, I'll tie it all up in the main typechecking function. That will allow me to start work on editor support immediatelly.

It will take some work to get to that point. That is what I should be my main mission. To turn my focus to that.

I still haven't even done the parsing for constraints. I'll leave that for later. Really, there is so much to do. It will take time to integrate it all and test it out.

Before I start work on instances, there is something that is bothering me.

```fs
    let generalize (forall_vars : Var list) (body : T) =
        let scope = !scope
        let generalized_metavars = ResizeArray()
        let rec replace_metavars x =
            let f = replace_metavars
            match x with
            | TyMetavar(_,{contents=Some x} & link) -> go x link f
            | TyMetavar(x, link) when scope = x.scope ->
                let x = {scope=x.scope; constraints=x.constraints; kind=kind_force x.kind; name=null}
                generalized_metavars.Add(x)
                TyVar x
```

I do not like this.

I think after all that metavars need to be linked to vars. Otherwise during validation and filling, I won't know what to replace them with. Also, the way this is done right now is not even correct.

10:50am.

```fs
    let generalize (forall_vars : Var list) (body : T) =
        let scope = !scope
        let h = System.Collections.Generic.HashSet()
        let generalized_metavars = ResizeArray()
        let rec replace_metavars x =
            let f = replace_metavars
            match x with
            | TyMetavar(_,{contents=Some x} & link) -> go x link f
            | TyMetavar(x, link) when scope = x.scope ->
                let v = TyVar {scope=x.scope; constraints=x.constraints; kind=kind_force x.kind; name=null}
                link := Some v
                replace_metavars v
            | TyVar v -> (if v.name = null && h.Add(v) then generalized_metavars.Add(v)); x
```

This should do well.

...Actually, this is hacky. Let me do it like...

```fs
    let generalize (forall_vars : Var list) (body : T) =
        let scope = !scope
        let h = System.Collections.Generic.HashSet()
        List.iter (h.Add >> ignore) forall_vars
        let generalized_metavars = ResizeArray()
        let rec replace_metavars x =
            let f = replace_metavars
            match x with
            | TyMetavar(_,{contents=Some x} & link) -> go x link f
            | TyMetavar(x, link) when scope = x.scope ->
                let v = TyVar {scope=x.scope; constraints=x.constraints; kind=kind_force x.kind; name=null}
                link := Some v
                replace_metavars v
            | TyVar v -> (if h.Add(v) then generalized_metavars.Add(v)); x
```

I do not want to make assumptions on ...

No, actually, let me just go with what I had. It is more efficient that way.

No need to get fancy.

At any rate, this will see me through.

Now the generalization should be correct.

Now, I think I should have a look at the nominal prototype instance. I'll leave the orphan checks for much later.

By the looks of things, it will be quite a while before I get to the prepass. There is just so much to do. But I need to take it in stride.

11:05am. Ok. The prototype instance is not the first thing I need to do. The first thing I need to do are top_inl and the recursive version of it.

I've somewhat ignored how I am going to deal with it. The easiest way to do it would be to just turn them into regular statements and pass that into the typechecker.

11:10am. If that is to work, one thing I would have to do is adjust the ranges when merging the blocks. No way around that.

11:30am. Had to take a short break. Let me just work on the non-recursive top level case. Forget other things.

11:50am.

```fs
let top_inl (r,name,expr,is_top_down) aux (top_env : Env) =
    if is_top_down then
        failwith ""
    else
        failwith ""
```

It is not coming to me. Also I've realized that in top level type definitions the ranges are incorrect as well.

12:05pm. Up to now, I've been so laser focused on finishing `infer` that I haven't spared much thought on how I am going to tie this all together.

And I went to sleep to damn late tonight so I feel very groggy even now.

12:20pm. I am thinking.

I know I said I would do the rest of the top level statements before moving on to editor support, but now that I am really focusing on the current step, I find that I do not really care about the top level statements. They aren't really the problem.

Now that the typechecker is in this advanced stage, instead what I have to do is start thinking about how to tie it up from the other end. There is nothing particularly challenging about the statements anymore.

Instead the real point of contention is that I've been loose about how exactly I am going to propagate information leading to the issues of dealing with ranges in mutually recursive statements and type definitions.

A certain idea has been coming to my mind. That would be to ignore the work that I've done thus far and instead when doing the first round of editor support, do a fake type on hover. Right now I do not dare run the typechecker because no doubt it has bugs in it.

But I can do a mock typechecker easily enough, just to enable editor support integration.

This will allow me to work through the editor support concerns without necessarily having to burden myself with the typechecker itself right off the bat.

If I have proper editor support when I start testing the real typechecker, debugging it will be a lot easier.

12:25pm. That is what I will do. I worked on the typechecker for a while, but now it is time to switch gears and go at it from the other end again.

Let me take a break here."